refactor: move incoming menu to incoming.jsx
