+++ botfront/imports/ui/components/nlu/activity/Activity.jsx
@@ -2,68 +2,30 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Meteor } from 'meteor/meteor';
-import { browserHistory, withRouter } from 'react-router';
+import { withRouter } from 'react-router';
 import { withTracker } from 'meteor/react-meteor-data';
 import { sortBy, uniq } from 'lodash';
 import moment from 'moment';
 import {
-    Tab, Message, Segment,
+    Message, Segment,
 } from 'semantic-ui-react';
 
 import { connect } from 'react-redux';
-import ActivityInsertions from './ActivityInsertions';
+
 import ActivityDataTable from './ActivityDataTable';
 import ActivityActions from './ActivityActions';
-import { Loading } from '../../utils/Utils';
 import { ActivityCollection } from '../../../../api/activity';
 import { NLUModels } from '../../../../api/nlu_model/nlu_model.collection';
 import { getPureIntents } from '../../../../api/nlu_model/nlu_model.utils';
 import { wrapMeteorCallback } from '../../utils/Errors';
-import ConversationBrowser from '../../conversations/ConversationsBrowser';
-import { updateIncomingPath } from '../../incoming/incoming.utils';
+
 import PrefixDropdown from '../../common/PrefixDropdown';
 
 class Activity extends React.Component {
-    // eslint-disable-next-line react/sort-comp
-    getDefaultState = () => ({
+    state = {
         filterFn: utterances => utterances,
-        activeTabIndex: undefined,
         sortType: 'newest',
-    });
-
-    state = this.getDefaultState();
-
-    createMenuItem = (name, index, tag = '', dataCy = null) => {
-        const { router } = this.props;
-        const regexp = / /g;
-        const urlId = name.toLowerCase().replace(regexp, '');
-        const url = updateIncomingPath({ ...router.params, tab: urlId });
-        return {
-            name,
-            content: `${name} ${tag || ''}`,
-            key: urlId,
-            'data-cy': `incoming-${dataCy || urlId}-tab`,
-            onClick: () => {
-                if (router.params.tab === urlId) return;
-                browserHistory.push({ pathname: url });
-                this.setState({ activeTabIndex: index });
-            },
-        };
-    }
-
-    getPanes = () => {
-        const {
-            model, instance, project, utterances,
-        } = this.props;
-        return [
-            { menuItem: this.createMenuItem('New Utterances', 0, `(${utterances.length})`, 'newutterances'), render: this.renderIncomingTab },
-            {
-                menuItem: this.createMenuItem('Conversations', 1),
-                render: () => <ConversationBrowser projectId={project._id} />,
-            },
-            { menuItem: this.createMenuItem('Populate', 2), render: () => <ActivityInsertions model={model} instance={instance} /> },
-        ];
-    }
+    };
 
     batchAdd = () => {
         const { modelId } = this.props;
@@ -85,7 +47,7 @@ class Activity extends React.Component {
         Meteor.call('activity.updateExamples', utterances, wrapMeteorCallback(callback));
     };
 
-    renderIncomingTab = () => {
+    render = () => {
         const {
             model: { _id: modelId },
             utterances,
@@ -144,29 +106,6 @@ class Activity extends React.Component {
             <Message success icon='check' header='Congratulations!' content='You are up to date' />
         );
     };
-
-    componentDidMount = () => {
-        const { router } = this.props;
-        const { activeTabIndex } = this.state;
-        const panes = this.getPanes();
-        const paneIndex = panes.findIndex(({ menuItem }) => menuItem.key === router.params.tab);
-        if (activeTabIndex === undefined) this.setState({ activeTabIndex: paneIndex >= 0 ? paneIndex : 0 });
-    }
-
-    render() {
-        const { ready } = this.props;
-        const { activeTabIndex } = this.state;
-        if (!ready) {
-            return <Loading loading={!ready} />;
-        }
-        return (
-            <Tab
-                activeIndex={activeTabIndex}
-                menu={{ pointing: true, secondary: true }}
-                panes={this.getPanes()}
-            />
-        );
-    }
 }
 
 Activity.propTypes = {
@@ -174,7 +113,6 @@ Activity.propTypes = {
     utterances: PropTypes.array.isRequired,
     entities: PropTypes.array.isRequired,
     intents: PropTypes.array.isRequired,
-    instance: PropTypes.object.isRequired,
     linkRender: PropTypes.func.isRequired,
     outDatedUtteranceIds: PropTypes.array.isRequired,
 };
